#19 method view container
