+++ src/js/pages/method-editor/components/summary-view.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
-import { Row, Col } from 'antd';
-import { translations } from '../../../translations';
+import { Method } from '../../../shared/method';
 
 
 export class SummaryView extends Component {
@@ -13,36 +12,8 @@ export class SummaryView extends Component {
     }
 
     render() {
-        const RawHTML = ({children, className = ''}) =>
-            <div className={className} dangerouslySetInnerHTML={{ __html: children.replace(/\n/g, '<br />')}} />;
-
-        const { seminarGoals, content } = this.props.method;
-
-
         return (
-            <div>
-                <Row>
-                    <Col span={24}>
-                        <h3>{this.state.method.title}</h3>
-                    </Col>
-                </Row>
-                <Row>
-                    <Col span={9} offset={3}>
-                        <h3>{translations.goals}</h3>
-                        <ul>
-                            {seminarGoals.map(item=> (<li key={item.id}>{item.name}</li>))}
-                        </ul>
-                    </Col>
-                    <Col span={9}>
-
-                    </Col>
-                </Row>
-                <Row>
-                    <Col span={18} offset={3}>
-                        <RawHTML>{content}</RawHTML>
-                    </Col>
-                </Row>
-            </div>
+            <Method loading={false} method={this.state.method}/>
         );
     }
 }
